Group logic related to providing Dropdown context and collection
